Add basic support for copying to the system clipboard via OSC 52
NOTE: this is an edited version of a commit created on 2021/04/22.
The comments about kitty are now out of date, but are being kept as
they originally were, in order to preserve some of the history. The
implementation and comments will be updated in a follow-up commit.

The new flags for the "copy" command are left undocumented for now,
which implies they may still receive backwards incompatible changes.