Remove kitty-specific OSC 52 workarounds
The OSC 52 "concat" extension in kitty was replaced by a standard
implementation in 0.22.0. This implementation imposes an 8MiB size
limit, but no longer requires the clipboard to be cleared before
sending the payload. It also no longer requires sending the payload
in chunks of <= 6138 bytes.

In kitty 0.23.1, the size limit was made configurable and was also
given a new, default value of 64MiB. Since we can't determine what
the limit is and since there's no longer any compatibility issues
with standard OSC 52, we can now just drop the kitty-specific
copy_text() handler.

NOTE: this means "copy -b" will no longer work properly with builds
of kitty older than 0.22.0. Since "copy -b" is still undocumented
and hasn't yet made it into a proper release, this can't really be
considered a "breaking change".

See also:

* https://sw.kovidgoyal.net/kitty/changelog
* https://github.com/kovidgoyal/kitty/commit/8f214c51c092d79e90fea25ec2e5db117af0054b
* https://github.com/kovidgoyal/kitty/commit/c00e945f6ee77e9f741304cc1586d82e418d4222
* https://sw.kovidgoyal.net/kitty/conf/#opt-kitty.clipboard_max_size