Add userdata pointer to CommandArgs struct
This allows us to pass an EditorState pointer to every command
handler function, which is necessary for eliminating the use of
global variables.

The userdata pointer is also stored in each CommandSet object, but
this is just for expedience of implementation. The details will be
cleaned up in later refactorings.

Note: the reason void pointers are used in some places instead of
EditorState pointers is to keep the code in src/command/* reusable
by other projects, with minimal (or no) changes.