Explicitly free EditorState struct on exit
Everything in this struct lives for the whole duration of the process,
so we were previously just allowing it to be freed implicitly on exit.
This works fine on any POSIX OS and also works well with LeakSanitizer.

However, some other memory checking tools (e.g. valgrind) don't have the
benefit of compiler instrumentation and can only check that the number
of allocations matches the number of frees. Even though LeakSanitizer is
generally faster and more convenient, we still want to maximize our
ability to use other tools. This commit thus adds most of the necessary
deallocation routines to a new free_editor_state() function.

This doesn't yet make the process 100% "valgrind clean" but it does the
majority of the work required to get there. The remainder can be done
later, either via explicit deallocation or valgrind suppressions.