Make relative_filename() work correctly with 2 identical arguments
Before this commit, if both arguments were the same string, the function
would crash due to an out-of-bounds read. This has never been a problem
in this codebase because the function in question has never been used
on directories and the "cwd" parameter is always a directory.

However, general-purpose utility functions like those in src/util/
should have well defined behavior for any input, so in this case
we should just return xstrdup(".").