Fix normalize_extended_keycode()
The "refine this" comment was stupid and written in a sleep deprived
state. The intention was to unify the way Alt+Shift+punctuation key
combos are interpreted under the legacy and kitty keyboard protocols,
but this naive "implementation" failed to do so and caused more
problems than it fixed.

The problem of one protocol producing "M->" and the other "M-S-." can
perhaps be tackled later, but it's harder to solve that it would seem.
For now, these kinds of key combos should be bound in both forms, to
maximize portability to different terminals with varying levels of
support for extended keyboard protocols.

Note: the test cases that have been commented out by this commit are
only applicable to "modifyOtherKeys" mode 2, which is never actually
activated by dte. Support for that (deeply flawed) mode is incomplete
anyway, so there's not much point in having tests for it.