Move cursor to last line for "show command" and "show search" buffers
Command and search histories are ordered such that the most recent
entry appears at the bottom of the list. The most recent entry is
usually the most relevant thing to show, so it makes sense to
position the cursor accordingly.

Reversing the order of entries was also considered, but it seems
sensible to keep them in the same order as they would be in the
on-disk history files.