Replace all "external" commands with a single "exec" command
The new command itself is left undocumented for now, which implies
it may still change, although it will remain capable of "emulating"
the old commands via the newly added, built-in aliases.

The only breaking change (intentionally) introduced by this commit
is the pipe-from "-s" flag (strip newline) being changed to "-n".
This was unavoidable and hopefully has limited (or no) impact for
most users.