Explicitly include <stdc-predef.h> in src/util/unicode.h, if it exists
This header is supposed to be included automatically by the compiler,
as a way to allow libc to pre-define things by default. However, even
though musl gained such a header[1] in 2016, it still doesn't appear
to be working as expected in some musl-targetted toolchains (like e.g.
the one provided by the Arch "musl" package[2]).

We use __STDC_ISO_10646__ as a way to determine whether the <wctype.h>
functions are sane, so maximizing the situations in which it's properly
defined helps to avoid unnecessarily using the ASCII-only fallbacks.

[1]: https://github.com/njlr/musl/commit/0f47367751a4baca80ef4f5ab11dbb9b6d7e8bdd
[2]: https://archlinux.org/packages/community/x86_64/musl/

Note: this commit also updates the git "commit-msg" hook, to permit
the URLs above.