Add xmemmem() function
This is intended to provide either a thin wrapper around memmem(3)
or a portable fallback implementation on platforms that are lacking
it.

For now though, it simply throws a compilation error if memmem(3)
isn't available. All the BSDs, Linux (both glibc and musl) and macOS
provide memmem(3) as a library function, so the lack of the fallback
isn't a major portability problem, so long as it materializes before
the next release.