Add "cursor" command to allow customizing the cursor style/color
This is left undocumented for now and is thus considered "unstable"
and possibly still subject to breaking changes.

One of the intended use cases for this feature is to allow the
cursor to be styled differently when the "overwrite" option is
in effect. For example, with the following config:

    cursor insert bar #00ff00
    cursor overwrite block #ff0000
    cursor cmdline bar

...pressing the Insert key (to toggle the "overwrite" option) should
cause the cursor to switch between a green bar and red block. When in
command/search mode, it should become a bar with the default color.