Temporarily disable bracketed paste mode, to work around a regression
Our use of xmemmem() in term_read_bracketed_paste() doesn't properly
handle partial end delimiters that get split on read boundaries.

This was rather a stupid mistake, but for now it's probably best to
just avoid enabling the mode until there's a proper fix, since the
consequences can make the editor unusable in the worst cases.