Add "toggle" command to search/command mode and remove "case"
The documentation for the "bind" command states the following for
command/search mode bindings:

> Most of these commands behave in a similar fashion to the normal
> mode commands of the same name.

However, the "case" command in normal mode was nothing like the "case"
command in the other modes. Adding "toggle" to other modes does away
with this discrepancy and allows using "toggle case-sensitive-search"
instead. This also makes the default M-c binding in search mode much
more self-explanatory.