Add -x command-line option, to allow loading extra rc files at startup
Files specified with this option are loaded after the default rc file
and multiple such files can be given.

If used as e.g.:

    dte -x rc1 -x rc2

...the following config files are loaded (in order):

1. $DTE_HOME/rc
2. $PWD/rc1
3. $PWD/rc2

To load extra rc files but exclude the default rc, the following can
be used instead:

    dte -R -x rc1 -x rc2

Note: this option is left undocumented for now and thus may still
receive breaking changes.