Fix bug in "exec" command with "buffer" output and non-buffer input
If the input comes from somewhere other than the buffer, the previous
logic for "-o buffer" was wrongly attempting to replace the text. This
was obviously wrong and could cause a crash in some cases, for example:

    tag main; open -t; exec -i msg -o buffer head -n1

The problem in this case is triggered by using "msg" as the input,
although any future, non-buffer inputs would be similarly affected.