Simplify definition of $(bashcompletiondir) in GNUmakefile
The bash-completion docs suggest using pkg-config to detect the correct
directory to install completion scripts to. However, this is very poorly
thought out, since it doesn't accomodate different values for $(prefix).
It potentially also leads to clashes between system packages and custom
builds, since the completion scripts are always being installed to the
same place, regardless of $(prefix). Furthermore, it makes the build
system depend upon both pkg-config itself and "bash-completion.pc".

macOS users may have to specify a custom location for bashcompletiondir
now, but that's not such a big deal compared to all the flaws of the
pkg-config method.