Handle EOF indicated as zero return from read() in handle_piped_data()
This can happen on platforms where poll() sets pollfd::revents to
POLLIN|POLLHUP on EOF. We can't immediately close the descriptor in
that case because it can happen either due to EOF or to indicate that
the other end of the pipe was closed when there's still buffered data
left to consume. Thus it falls to read() returning zero to
differentiate the two.

Many thanks to Kevin Hovsäter for his excellent help in debugging and
testing this!

See also:

* https://www.greenend.org.uk/rjk/tech/poll.html