Improve logic for when to freopen() /dev/tty in init_std_fds()
Bad things happen if we try to use a terminal other than the
controlling terminal, as would happen when running e.g.:

    setsid dte

Therefore, init_std_fds() should check whether descriptors 0-2 are
the controlling terminal, not just any terminal. In the case of the
above example, this would lead to a failed freopen() call and then
exiting with an error message, which is exactly what we want.