Log warning if syntax command uses current state name as destination
In dte-syntax(5) files, the current state can be conveniently referred
to by using "this" as a destination argument (e.g. `char a-z this`).
This involves much less overhead since there's no lookup required.

Using the full name of the current state instead of using "this"
should probably be considered an error, but for now we just log a
warning, since we don't want to suddenly break existing syntaxes (as
would happen if error_msg() were used instead).