Set SIGXFSZ signal disposition to SIG_IGN at startup
This signal is sent if the process tries to create a file larger than
the RLIMIT_FSIZE resource limit allows. If the signal disposition is set
to ignored (SIG_IGN), the library function in question returns EFBIG and
allows us to handle it as a normal error instead of terminating.

See also:

* https://pubs.opengroup.org/onlinepubs/9699919799/functions/getrlimit.html