Extend "line" command to accept a line[,column] position argument
Instead of only accepting a line number, as was the case before.

The format line[:column] (with a colon instead of a comma) is also
accepted, which mirrors the +line[:column] command-line arguments
mentioned in the dte(1) man page.

This change is left undocumented for now, which implies it may still
receive breaking changes.