Silence pandoc warning about "markdown_github" being deprecated
This deprecation notice is extremely dumb, since the suggested
replacement ("gfm") is not at all viable and doesn't support any
of the commonly used extensions.

If this mode is eventually removed from pandoc, the chosen
replacement will absolutely *not* be another pandoc format,
it'll be an entirely different tool that doesn't constantly
break compatibility.