Simplify Markdown formatting used in docs/dte.md
Having emphasis spans right next to each other with no spacing is
interpreted differently by different Markdown processors and seems
to cause problems in general. It's not essential to have literal
characters bolded in dte(1), so for the sake of simplicity and
avoiding issues we can simply drop the ambiguous strong emphasis.