Add proper handling for all modifiers in kitty keyboard protocol
This means Super and Hyper are now supported in key bindings and
NumLock/CapsLock are also ignored, which fixes an issue causing
bindings to not work when either or both were enabled.

The new Super/Hyper support is left undocumented for now, which
implies it may still receive breaking changes.

See also:

* https://sw.kovidgoyal.net/kitty/keyboard-protocol/#modifiers
* https://www.gnu.org/software/emacs/manual/html_node/efaq/Binding-combinations-of-modifiers-and-function-keys.html