Use full `timespec` struct instead of `time_t` for FileInfo::mtime
This gives nanosecond precision to file modification times and also
allows the full time to be included in `show buffer` output.

Note: for some reason macOS still doesn't seem to allow accessing the
`stat` modification `timespec` as `stat::st_mtim`, even though it's a
mandatory part of POSIX 2008, so a getter function is used as an aid
to portability.