Add built-in `$MSGPOS` variable
This expands to the position of the current message (as would be
specified with e.g. `msg 42`) or "1" if there are no messages.
Defaulting to "1" is done as a convenience for typical use cases,
although this behavior can be revised later if it turns out to be
problematic.

Example usage:

    git-grep example
    exec -i msg -o msg fzf --reverse --bind 'start:pos('$MSGPOS')'

Note: this feature has been left undocumented for now, which implies
it may still recieve breaking changes in the future.