Add `-E` flag to built-in `git-grep` alias, to enable extended regexps
This is technically a breaking change for anyone who has been using
`git-grep`, but hopefully it's not too disruptive. The previous behavior
can be restored by using `git-grep -G` or by overwriting the alias with
the old definition.

This is being done now because it makes the `git-grep` pattern format
match the one used by the `replace` command and because Basic Regular
Expressions (BRE) are unnecessarily verbose for typical usage.