Add optional argument to `copy` command, to allow copying arbitrary text
This allows the copied text to be specified via manual entry in command
mode or perhaps via the expansion of variables in bindings and aliases.

It also allows passing arbitrary text to the system clipboard using the
OSC 52 clipboard protocol, which currently isn't straight forward to do
with e.g. the `exec` command and external tools.

Note: this feature has been left undocumented for now, which implies
it may still recieve breaking changes in the future.