Mention ECMA-48 reference in src/terminal/parse.c
Section 5.4 of ECMA-48 concerns the format of control sequences, as
sent from the client to the terminal. Even though communication in
the opposite direction doesn't appear to be specified anywhere, the
same format is used in practice. The details mentioned there are
useful for understanding how to write a general purpose parser, as
opposed to just doing naive string matching.