Fix regression in cursor positioning after negative `shift` command
More specifically, the cursor always jumping to the end of the line
after a command like e.g. `shift -- -1`, instead of remaining on the
same character.

This regression seems to have occured due to multiple changes between
the v1.8 and v1.9 releases and is caused by integer overflow in an
expression with mixed types (and thus implicit conversions):

    long x = view_get_preferred_x(view) + (count * width);

Example reproducer:

    open -t
    set -l indent-width 4 expand-tab true emulate-tab true
    insert '        1 2 3 4'
    repeat 2 right
    shift -- -1