Fix definition of `MAKE_S` variable in `mk/util.mk`
The removed part of this definition was causing `make portable` to
be unexpectedly silent when building the binary. This was due to the
custom `MAKEFLAGS` for that target containing an unrelated `-s` flag
(for `LDFLAGS`), which was erronenously being interpreted as make's
own `-s` flag.

This new definition for `MAKE_S` appears to work fine with GNU Make
4.0, which is the oldest version the dte build system supports.