Fix recently introduced bug in `activate_current_message_save()`
In commit 52e2a522a0eed536 `file_location_go()` was changed to return
the newly active `View`, unless either the file couldn't be opened or
`FileLocation::pattern` (if any) wasn't found. The returned `View` was
then used in `activate_current_message_save()` as part of the condition
to determine whether or not to call `bookmark_push()`.

One circumstance where this adjustment to the condition changed the
existing behavior was in the case where `file_location_go()` changed
the current view but then returned `NULL` anyway, due to `search_tag()`
failing to find the pattern. A bookmark was previously still pushed in
that case, until the offending commit broke it.

Any time the current view or cursor position changes in response to a
command like `tag`, a bookmark should always be pushed, even for unusual
edge cases. This commit thus restores the previous behavior by reverting
the relevant parts of commit 52e2a522a0eed536.