Allow most flags from normal mode commands in command/search mode
Most of the flags that aren't needed in command/search mode are now
simply ignored, as a convenience for binding in multiple modes with
a single `bind` command. The bulk of the exceptions to this are `-c`
and `-l` selection flags, which wouldn't really make sense to ignore
and most likely aren't going to be bound in multiple modes anyway.