Adjust `DEBUG` usage, so that sources don't depend on the build system
Source files can still be *affected* by passing e.g. `-DDEBUG=2` to
the compiler, but the fallback definition (i.e. `DEBUG=0`) now lives
in `src/util/debug.h`.

This allows files to be copied to other projects without needing to be
separated from the build system. In this case, only a trivial change
is required, but having the 2 components be easily separable should be
maintained as a general principle, since the code under `src/util/` is
explicitly intended to be reusable.