Add `-l` flag to left/right/bol/eol/word-bwd/word-fwd commands
To allow them to extend the selection while keeping it line oriented.

There's no particular reason not to support this kind of usage and it
may be more useful now that the editor supports custom input modes.

Note that these new flags haven't been added to the `dterc(5)` man page
for now, but they can nevertheless be considered stable. Documentation
will be added before the next release.