Add `O_NOCTTY` to `open()` flags automatically in `xopen()`
Acquiring a new controlling terminal is almost never a desired outcome
of opening a file, but it can be done by accident in some situations
(although it's either impossible or very unlikely in this codebase,
depending on the platform).

Given how rare it is to specifically want to omit `O_NOCTTY` (i.e.
never in this codebase), it may as well always be added to the flags
by `xopen()`.

Note that the above reasoning factors in the fact that code under
`src/util/` is specifically selected/implemented with use in other
codebases in mind.

See also:

* https://pubs.opengroup.org/onlinepubs/9699919799/functions/open.html
* https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap11.html#tag_11_01_03