Add `dte -S` command-line option, for verbose linting of syntax files
This is almost the same as the `-s` option, but it also warns about
redundant emit-name arguments for conditionals and the `eat` command.
More "verbose" checks may also be added later.

Note: this feature has been left undocumented for now, which implies
it may still receive breaking changes in the future.