Silence `-Warray-bounds` false positive in `get_delim_str()` tests
GCC tries to be too clever about this and uses a bunch of heuristics
that find non-issues and then report them in a misleading way. The
idea behind the way the code was originally written was to trigger
ASan if real bugs arise, instead of certain edge cases potentially
going unnoticed.