Add `-S` flag to `delete-line`, to limit scope of deletion to selections
Normally running `delete-line` when there's a character selection will
delete every line "touched" by the selection (even characters that fall
outside the selection). When `delete-line -S` is used in conjuction with
selections, the command now behaves identically to `delete`.

This may seem counter-intuitive, given the name, but some people may
wish to bind a key combo to the command and use it both for single
(unselected) lines and selections, with a preference for the `-S`
behavior on selections.

Note: this feature has been left undocumented for now, which implies
it may still receive breaking changes in the future.