Move `search` command to "movement commands" section of `dterc(5)`
It ended up in the "editor commands" section when the manual was
re-organized, probably because it wasn't clear to me at the time
which category to put it in. Upon reflection, it's fairly obvious
that its main purpose is to jump to specific parts of the buffer
(thus moving the cursor).