Be more selective about sending DECRQM query sequences to the terminal
The FreeBSD 14.0 console seems to parse these sequences incorrectly, by
interpreting the intermediate `$` as the terminator instead of the final
`p`. This is contrary to the ECMA-48 specification of CSI sequences and
results in a stray "p" being drawn on screen at startup (until input
triggers an update).

Since the FreeBSD console sets `TERM=xterm` by default, that's what we
use to determine whether to avoid sending the queries. In the future, it
may be possible to switch to a more precise condition for this, although
the consequence of some other terminals not having `TFLAG_SYNC` set via
detection is only a minor thing.

For quick reference, the sequence in question is:

    \033[?2026$p

See also:

* https://ecma-international.org/publications-and-standards/standards/ecma-48/
* https://vt100.net/emu/dec_ansi_parser