Fix `path_absolute()` test failure on Solaris
Apparently the canonical path for `/dev/null` on Solaris resolves to
`/devices/pseudo/mm@0:null`, which breaks the assumption made by this
test case. To fix this, we simply use `/dev/` instead of `/dev/null`,
which still exercises the code path we're interesting in testing.

Many thanks to Mike (@tankf33der) for reporting this issue!