Avoid returning `KEY_IGNORE` from `term_read_key()`
Some callers of this function aren't expecting to handle it and until
recently never saw it as a return value. It's essentially meaningless
except within `term_read_key()` and the various lower level functions
it calls, so there's no reason to ever return it or make higher level
callers handle it.

This fixes an issue where opening a file larger than `filesize-limit`
at startup would cause an error to be shown in the status line, but
then be immediately cleared by query responses. The condition in
`main_loop()` only checks for `KEY_NONE` and anything else causes the
loop to proceed as if real keyboard input was received, thus clearing
the error message.

Simple reproducer:

    kitty dte 300MiB.txt