Avoid using `timezone` or `daylight` in `timespec_to_str()` tests
The former appears to be a function[^1] on FreeBSD and the latter isn't
declared at all in `<time.h>`.

For now, we simply switch to running the tests that were previously
guarded by these 2 values unconditionally. If this causes failures on
some platforms, then they can just be dropped later.

See also:

* https://pubs.opengroup.org/onlinepubs/9699919799/functions/timezone.html
* https://man.freebsd.org/cgi/man.cgi?query=timezone&sektion=3
* https://man7.org/linux/man-pages/man3/tzset.3.html#HISTORY

[^1]: As opposed to an `extern long`, as it is on Linux and in the Open
Group XSI spec.