Rename constants for `ExitCode` enum in `src/util/exitcode.h`
Apparently `EX_OK` clashes with an indentifier from a system header
when compiling on IRIX with SGUG RSE's GCC:

    In file included from src/editor.c:27:
    src/util/exitcode.h:24: warning: "EX_OK" redefined
    24 | #define EX_OK 0
    |
    In file included from /usr/sgug/include/libdicl-0.1/unistd.h:5,
    from /usr/sgug/include/libdicl-0.1/repl_select.h:4,
    from /usr/sgug/include/libdicl-0.1/sys/time.h:9,
    from /usr/include/sys/stat.h:25,
    from /usr/sgug/include/libdicl-0.1/sys/stat.h:5,
    from src/editor.c:7:
    /usr/sgug/lib/gcc/mips-sgi-irix6.5/9/include-fixed/unistd.h:51:
    note: this is the location of the previous definition
    51 | #define EX_OK 020 /* Test for Regular, executable file */

Although the `exitcode.h` header appears to have been edited here, the
clashing identifier still poses a potential problem. To avoid the whole
issue, we can simply rename the constants, which is perhaps a good idea
anyway for other reasons.

Thanks to chulofiasco of the IRIX Network Forums for providing this
build log.