Replace `void *userdata` params in `src/command/*` with `EditorState *e`
This fixes some `-fsanitize=function` errors by making the declared and
called function pointer types match. Not doing so would be UB according
to a strict interpretation of ISO C, although in practice every platform
we support does the sane/obvious thing and the UB is just pedantry in
the context of POSIX requirements.

Even so, this avoids the need for casting and/or sanitizer suppressions
and as a side benefit also makes the code slightly clearer.

The original reason for using `void* userdata` parameters was to make
the code more amenable to being turned into a library, but I no longer
consider that a useful goal for this part of the codebase anyway.