Call `FreeFunction` callbacks via `no_sanitize` wrapper function
Clang 17 `-fsanitize=undefined` enables `-fsanitize=function`,
which produces the following sanitizer errors:

    src/util/ptr-array.c:65:9: runtime error: call to function
    free_filetype_entry through pointer to incorrect function type
    'void (*)(void *)'

    src/util/hashmap.c:231:13: runtime error: call to function
    free_syntax_cb through pointer to incorrect function type
    'void (*)(void *)'

As mentioned in commit 79d91e600fe0, this type of error is mostly
pedantry and isn't a problem in practice on any POSIX platform we
support. Perhaps a different method can be used later, but for now
this avoids showing known/deliberate "errors" when compiled with
Clang `-fsanitize=undefined`.