Use cc builtins in `src/util/bit.h` if `HAS_BUILTIN()` indicates support
Clang and recent versions of GCC support both the `HAS_BUILTIN()` checks
and the `GNUC_AT_LEAST()` checks and older versions of both also support
the latter, so this commit changes nothing for those 2 compilers.

However, `__has_builtin()` is much easier for alternative compilers to
support, since it can be done in a piecemeal fashion. Thus, this commit
is done "in principle" for the benefit of such compilers (now or in the
future).