Make length-bounded string functions accept `NULL` arguments when `n=0`
As far as I'm aware, none of our uses of these functions run into real
problems related to this. However, passing `NULL` arguments to library
functions that aren't expecting them (e.g. `memcmp()`) is technically
UB, even when the length bound is zero.

Since there's a solution to this pitfall that (even in most accidental
cases) does the right thing and adds little to no overhead, it makes
sense to use it in dte's string functions, where applicable.

See also:

* https://www.open-std.org/jtc1/sc22/wg14/www/docs/n3261.pdf