Don't error out at startup if `$TERM` is unset
The editor already assumes a minimal set of ECMA-48 control functions
are supported and all the extended features in the built-in `terms[]`
"database" can now be detected via querying the terminal. We also
already error out in cases where there's no controlling terminal.

Since `$TERM` is a variable that's fundamentally tied to terminfo(5)
and since we never actually use terminfo, there's no good reason to
require that it's set in order to successfully start the editor.