Remove empty/unset `$TERM` tests from `test/check-opts.sh`
These tests fail if the script is run in a context without a controlling
terminal, which makes most of the GitLab CI and GitHub Actions jobs
fail.

There's really no need to test this any more and it was just mindlessly
converted in the previous commit when it should have been removed.