Simplify path taken to enter `showkey_loop()` from `main()`
This entirely does away with logging when using `dte -K`, but that's
probably a good thing, since such logs serve no particular purpose and
are less complete than regular logs.

The reason this is being done is because the previous state of things
was making refactoring code in `main()` more awkward than necessary.