Add comment to `test/terminal.c` regarding `CSI R` ambiguity
We never emit queries that produce a CPR (Cursor Position Report)
response, so it's reasonably safe to assume that `CSI R` with 2
parameters is an input event for the F3 key. However it may be useful
to start making use of cursor position reports later, so this comment
serves to help avoid confusion, if this test starts failing.

See also:

* https://sw.kovidgoyal.net/kitty/keyboard-protocol/#:~:text=CSI%20R
* https://invisible-island.net/xterm/ctlseqs/ctlseqs.html#:~:text=Device%20Status%20Report%20(DSR)
* https://vt100.net/docs/vt510-rm/DSR-CPR.html
* https://vt100.net/docs/vt510-rm/CPR.html