Remove Clang `counted_by` attribute from `HashSetEntry::str`
The flexible array member is explicitly null-terminated and an extra
byte is allocated to accommodate it but the terminator isn't included
in the `str_len` member, which leads to the following runtime error:

    $ make check CC=clang
    src/util/hashset.c:110:5: runtime error: index 15 out of bounds
    for type 'char[]'

This is simply a misuse of the attribute, rather than an indication
of a real problem in the code. It seems there's currently no way to
use the attribute and instruct it to treat the length member as being
one byte short of the actual allocation, so for now we simple avoid
using it.