Add `log_level_debug_enabled()` and `log_level_trace_enabled()`
So that would-be callers of `log_level_enabled()` don't have to check
the value of `DEBUG` to get proper dead code elimination, when one or
both of these log levels are disabled at compile-time.

This is done for the sake of moving all such `DEBUG`-checking logic
to a single location.