Disable `modifyOtherKeys` mode before upgrading to Kitty Keyboard Mode
When support for the latter is detected via dynamic querying.

This seems like the correct thing to do, because in this case we
previously emitted the sequence to enable `modifyOtherKeys=1` and
when we disable Kitty keyboard mode[^1] on exit, `modifyOtherKeys`
mode may have remained active (and merely superceded/masked by the
Kitty mode).

This all depends on the terminal and I haven't yet tested any of
those that support both modes. However, this commit is unlikely to
do any harm and is done for "correctness in principle".

[^1]: by calling `term_restore_private_modes()`