Improve escaping of backslash characters in `docs/pdman.lua`
Apparently there are some scenarios where `\\` doesn't work as
expected and, in any case, `\e` is the standard escape for this.

This commit also adds table fields for escaping for backticks and
double quotes. There appears to be no existing instances of the
former and the latter is (for now) simply "escaped" to itself,
since according to `groff_char(7)` it currently maps to itself
anyway.