Optimize and clean up `expand_word()`
This avoids the need to call `xrealloc()` immediately after `xmalloc()`,
removes the hacky `view_get_selection()` function and also improves the
behavior for the following series of commands:

    set select-cursor-char false
    insert xyz
    right -c
    left -c
    insert $WORD

Previously this example would replace the zero-length selection with
the word under the cursor (e.g. "xyz"), whereas `$WORD` now expands
to nothing in this case and the result is as if `unselect` was used.