Bind `M-escape` the same way as `escape` in `config/binding/default`
Most terminals have some quirky behavior when the Escape key is held
(due to escape sequence ambiguity and the related timing issues), but
rxvt-unicode goes one step further and makes *all* subsequent Escape
key presses produce `M-escape` (until another key is pressed to reset
the stuck state). I'm not sure yet what causes this, but it's also
reproducible in dex and my suspicion is that it's a bug in rxvt.

In any case, binding `M-escape` like this is harmless and may provide
some marginal (albeit rarely needed) benefit in rxvt.