Add `-q` flag to `bind` command, to allow silently ignoring invalid keys
The intended use case for this is to avoid situations like the one
mentioned in commit 706d44666c804053.

If a similar situation should arise in the future, new bindings can be
added to the user `rc` file with `bind -q` and pre-existing support for
the `-q` flag then allows error messages for deliberate "errors" to be
ignored.

For example:

    # For old versions
    bind C-h erase
    bind C-? erase
    # For new versions; ignored by old versions supporting `bind -q`
    bind -q backspace erase

Note: this feature has been left undocumented for now, which implies it
may still receive breaking changes in the future.