Work around Backspace bug in WezTerm's Kitty keyboard implementation
If WezTerm's `enable_kitty_keyboard` option is set to `true`, e.g.
with the following command:

    # WARNING: this will overwrite your config
    echo 'return {enable_kitty_keyboard = true}' > ~/.config/wezterm/wezterm.lua

...then the following encodings can be observed, when running `dte -K`
and pressing various Backspace key combos:

    $ dte -K
    Press any key combination, or use Ctrl+D to exit
      backspace
      C-^H
      M-^H
      S-^H
      C-M-^H
      C-S-^H
      M-S-^H
      C-M-S-^H

These strings correspond to literal `BS` control characters ending
up in the `KeyCode` values returned by `normalize_csi_u_keycode()`.
While this does represent an issue that should be addressed in dte,
the underlying cause seems to be that WezTerm becomes "confused"
when `modifyOtherKeys=2` mode is toggled on and off and Kitty
Keyboard mode is then subsequently entered:

    $ printf "\033[>4;2m\033[>4m\033[>5u"
    $ cat -v
    ^[[127::8;5u

The sequence produced here (for Ctrl+Backspace) doesn't conform to the
Kitty protocol. In a different terminal that supports both protocols
(foot), the sequence produced for the same key combo is as expected:

    $ printf "\033[>4;2m\033[>4m\033[>5u"
    $ cat -v
    ^[[127;5u

Note that these same mode requests are in fact emitted by dte in
WezTerm's case, due to the way `TFLAG_KITTY_KEYBOARD` is handled in
`handle_query_reply()`. This is by design.