Fix handling of cursor position for some uses of `delete-line -S`
When `delete-line -S` was used on a character selection with the cursor
beyond the anchor point (e.g. as created by `repeat 10 right -c`), the
cursor was ending up at the same position (or at eol), instead of moving
left by the number of deleted characters (i.e. to the anchor point).