Don't unhide cursor in `update_screen()` if `UPDATE_DIALOG` flag is set
This fixes an issue where the cursor was visible inside the dialog,
when positioned under it prior to `dialog_prompt()` being called.
It was also visible in other cases too, but these were much less
noticeable/ugly.

Doing this also avoids unnecessary work in general, since the cursor
will always be restored and shown after `dialog_prompt()` returns.

The issue in question was introduced in commit 9761a6872466b1c4.