Add "headless" editor mode
This makes the following commands fail with an error message, in
cases where they'd usually prompt for interactive input:

* `close -p`
* `wclose -p`
* `quit -p`
* `replace -c`

It also disables the `suspend` command entirely and makes `exec`
and `compile` always operate in "quiet" mode (i.e. as if the `-s`
flag is always used).

For now, this mode is only used by the test runner.