Automatically disable syntax highlighting in files with very long lines
When the cursor is placed towards the end of a long line, moving
horizontally or inserting text can exhibit poor responsiveness, due to
the way in which syntax highlighting is currently handled. This issue
will be addressed properly at some point in the future, but for now we
simply work around it.