Clean up outdated code for special case in `window_open_buffer()`
The example in the removed comment hasn't been relevant (i.e. a
special case to this function) since we started using `realpath(3)`
in `path_absolute()`. `realpath(3)` simply resolves it to the absolute
path of the symlink (which is the closest thing to an absolute and
canonical path we can get in that case), regardless of whether it's
specified as a relative or absolute path.