Make `scroll-up`/`scroll-down` move cursor at margins (even if non-zero)
Also add a new `-M` flag to both commands, which makes them never
move the cursor (thus becoming no-ops when at the scroll margins).

Previously there was an inconsistency where a non-zero `scroll-margin`
would produce the no-op behavior, whereas with `set scroll-margin 0`
the cursor would always be moved when at the margins. This commit makes
these 2 commands behave the same way, regardless of the `scroll-margin`
value.

The `scroll-pgup` and `scroll-pgdown` commands will be amended in a
similar fashion, in a follow-up commit.

Note: this feature has been left undocumented for now, which implies
it may still receive breaking changes in the future.