Avoid checking the output of `dte -h` in command-line options tests
The value of `argv[0]` (as included in the help output) is not under
our control and on Termux it's an absolute path rather than just "dte".
This was causing the `check_str` test to fail. Since the purpose of this
test is mostly ASan/UBSan coverage, we can just run it without checking
the string.