Add `bol -r`, for jumping back and forth between BOL and end of indent
This is much like the existing `bol -t`, but moving to column 1 in all
cases except when already there.

Note: this feature has been left undocumented for now, which implies
it may still receive breaking changes in the future. In particular,
the flag may be changed from `-r` to a different letter.